Improve state caching.
The cache now tracks the age of entires, and deletes any that have not
been touched recently. MAX_IDLE_LIFETIME controls this.

Note that, if we could reliably hook into the moment just before the
session is saved, we could reduce MAX_IDLE_LIFETIME by 1.

There are tricky trade-offs here from a performance point-of-view.
* It is important to keep the session small. Therefore we should purge
entires as soon as possible.
* Some question engines, e.g. OpenMark, keep in-memory state, so it is
important to stop question sessions once they are no longer used.
* Other question engines are slow. E.g. ounit, are quite slow. They would
benefit from a longer cache life-time.

(It would be conceivalbe, but more complex, to allow lifetime to be
configured per-engine.)