Improve Opaque to integrate with the enhanced 2.7 quiz reports #10695
1. Response history under a question indicates which steps were a
submit, and which just a save.

2. Quiz admin -> Results -> Responses can show the response for all and
first tries, in the cryptic form used in the response history. Last try
is still shown in the form sent back by answerLine.

3. Response summary shown consistently in alphabetical order.

4. Two questions in a quiz that should use the same randomisation, now
will. That is #4179 should now be fixed.

5. Because of that in question preview, you can select a question
variant in the question preview pop-up. Note that this is not the same
as OpenMark's 'Fixed variant'.

6. In the quiz statistics there should now be a breakdown by variant,
but for this to work nicely, it relies on questions to send back a
questionLine including "variant = NNN". That is matched
case-insensitively, with any amout of whitespace (including none) around
the = sign. Otherwise the question is assumed not to have variants.